@tschottdorf 2017-07-19
My main concern here is that while making 1PC transactions really fast is great, it's going to make it even more surprising to folks who add an index and find that performance drops significantly as they don't get the 1PC path any more.