@spencerkimball 2017-07-20
I dislike the idea of optimizing for this case, which I suspect is middling unusual, at the expense of possible starvation. I do like some of the suggestions in the PR description though. In particular, it would be interesting to augment the `EndTransaction` with the keys read during the txn (if 3 or less and serializable let's say). Then, we could either commit despite the write too old error if snapshot or number of keys read is 0. If serializable, and we have the list of keys read, we could ping those keys with a new RPC that instead of reading, verifies and updates the timestamp cache entry. The verification and update is basically a compare and set on the entry's timestamp. If equal to the expected `txn.Timestamp`, would update the entry to the txn commit timestamp; if not equal, then would return an error, leading to a txn rety.

This is definitely non-trivial but depending on how many read keys we're willing to track, could short circuit a lot of serializable retries.