@bdarnell 2017-07-20
> I dislike the idea of optimizing for this case, which I suspect is middling unusual, at the expense of possible starvation.

I agree.

---

Review status: 0 of 1 files reviewed at latest revision, all discussions resolved, some commit checks failed.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/17121#-:-KpW8_gZQ3YohFty2tkw:bqhpsfx)*
<!-- Sent from Reviewable.io -->