@petermattis 2017-07-20
@tschottdorf My uncontended numbers are right in line with yours. Wonder why it was faster on May 9.

```
~ kv --cycle-length 1 --concurrency 1 --duration 30s
...
_elapsed___errors____________ops___ops/sec(cum)__p95(ms)__p99(ms)_pMax(ms)_____seq(begin/end)
   30.0s        0          23091          769.6      1.4      1.7     10.5              0 / 0
```