@petermattis 2017-07-20
Tested with ce9d95dd646ecb4c489233c282cc9484efa0b904 (a commit from May 9th):

```
~ kv --cycle-length 1 --concurrency 1 --duration 30s
...
_elapsed___errors____________ops___ops/sec(cum)__p95(ms)__p99(ms)_pMax(ms)_____seq(begin/end)
   30.0s        0          61102         2036.7      0.7      2.0     11.5              0 / 0
```

That's worth investigating!