@tschottdorf 2017-07-20
Wowzers, someone broke something. Will take a look next week (unless you beat me to it, @petermattis).

As for the PR, looks like we all agree that optimizing the 1PC path more is not a good idea. I'll write up some of the learnings in an issue and abandon this PR.