@petermattis 2017-07-21
I'm already bisecting. Will know later tonight.

On Thu, Jul 20, 2017 at 7:52 PM Tobias Schottdorf <notifications@github.com>
wrote:

> Wowzers, someone broke something. Will take a look next week (unless you
> beat me to it, @petermattis <https://github.com/petermattis>).
>
> As for the PR, looks like we all agree that optimizing the 1PC path more
> is not a good idea. I'll write up some of the learnings in an issue and
> abandon this PR.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/cockroachdb/cockroach/pull/17121#issuecomment-316862092>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AF6f97ci6_77O_0bXZzW_Saany2_V8n_ks5sP-gegaJpZM4OdQL->
> .
>