@petermattis 2017-07-21
Ah, it's 82cbb49ff7e1db6e1b01d96f1100954b599b6c5f. Specifically, the default 1ms value for `rocksdb.min_wal_sync_interval`. So this result is to be expected when there is no concurrency.