@tschottdorf 2017-07-21
Hmm, ok. Is there a lightweight way to make the `1ms` interval adaptive? Single-threaded inserts aren't found in many real-world workloads, but they're probably the first stupid benchmark people run.